// packages/b_declarations/src/properties/font-variant-ligatures/types.ts
// b_path:: packages/b_declarations/src/properties/font-variant-ligatures/types.ts

export type FontVariantLigaturesIR = {
  kind: "keyword";
  value:
    | "normal"
    | "none"
    | "common-ligatures"
    | "no-common-ligatures"
    | "discretionary-ligatures"
    | "no-discretionary-ligatures"
    | "historical-ligatures"
    | "no-historical-ligatures"
    | "contextual"
    | "no-contextual"
    | "inherit"
    | "initial"
    | "revert"
    | "revert-layer"
    | "unset";
};